shared/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "shared-llm"
version = "0.1.0"
description = "Unified LLM client with tier-based routing for AI Projects workspace"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
]

[project.optional-dependencies]
anthropic = ["anthropic"]
gemini = ["google-genai"]
openai = ["openai"]
all = ["anthropic", "google-genai", "openai"]

[tool.setuptools.packages.find]
where = ["."]
include = ["llm*"]
